Input Manager created
cranhky now supports a basic input system

A Input Manager struct created, that updates a std::bitset.
Change Texture Manager, now saves already loaded textures to an unordered_map.

+++ include/headers/ecs/system-manager.h
@@ -26,5 +26,5 @@ public:
     /**
      * Executes the underlying logic of each system at each frame
      */
-    void tick(int deltaTime);
+    void tick(float deltaTime);
 };
